Update README to have actual content; add guidelines
So far, the README file was empty.
There is now an actual first version of it, although most of the
"introduction" part is still pretty short. Consider it WIP.
The guidelines for importing projects, which we talked about during
the meetings, has now also been added to it.

I also deleted the "Guides & Help" folder now, because it isn't
really needed in this repository, and was kind of distracting.