Add materials for our PG color theme colors
We discussed this when we talked about the theme.
These materials represent the colors of our theme. The material
colors itself are not final (and maybe even the names are wrong,
we kind of have a hard time remembering) yet.

We are supposed to only be using these materials when creating
objects, so that we can later just change the material colors, and
all objects are updated.

Currently, there are some materials for normal objects in scenes,
and a few for 2D text elements (which need a special kind of
material). The 2D text materials are not yet completed.