Create a template room to build more rooms from
This room is, as the title said, supposed to be a help for building
more rooms from it.
It contains "debugging" objects which represent the VR Lab, so that
we can build our rooms with it in mind.

The room is not yet finished. I am not sure yet if the sides are
correct, and how it looks in VR.