Fix Y-position of text in the template room, add new assets
The text was floating too high above the ground. Now it should be
very close to it, but without introducing Z-fighting at any distance
we are going to be viewing everything from.

The new assets have been placed in the scene, however, it is likely
that they are still scaled weirdly, and are generally placed weird.
I will need to go to the VR Lab to finalize this, and also to
properly place the Lighthouse cameras and create their "poles".