Basic Room: Verify sizes, add VRLab walls, move & resize dummy
It is now confirmed that the "Tracking Room" in the "Basic Room"
template scene is correctly sized and positioned.

Also, correctly (I hope) positioned and sized walls for the VRLab
itself have been added as well.

The human dummy has been made a little less tall (since not everyone
is over 190cm tall) and moved to the side, since it was standing
right at the scene origin and generally kind of in the way.