Create ScriptableObject based keycodes, with custom inspector
This is the first step of creating a system for managing all the
keycodes we have within our game. The keycodes are a central part of
it, since they connect all our isolated scenes together, in
conjunction with the portal and the keypad.

The keycodes are ScriptableObjects because this way, they can be
directly added to GameObject components everywhere, providing a way
of (hopefully, in the end) doing most of the keycode and scene
connection logic within the Unity Inspector during "edit time", in
a (fairly) safe way.
This also makes it very easy to create new keycodes, and to make
sure that they are correct.