Add all my utility code to the fusion project
These are a lot of files at this point.
This is pretty much all I have. The namespaces for most of these
should be up to date, as well as other code-related stuff. Most of
this code is used somewhere, some in only a handful places, other
stuff in a lot of places.

I would have liked to have some more time to polish and continue
some of this (especially the Editor GUI component system), but
things are how they are, and I need this code now.

Yare yare daze.