Add attribute to Keycode to make it "discoverable", Editor additions
The idea is that other scripts can directly reference any keycode,
by just dragging it in in the Inspector, and the scripts can then
set the keycode to "discovered" anytime they want.
(A discovered keycode is one that the player can view and use in
the game.)

The Editor for the Keycode class has also been improved and cleaned
up a bit.