Create a (new) "KeycodesList" class
This class is a SerializedObject, and holds a list of all keycodes
(represented by SerializedObject asset files, too) mapped to scene
files.
The class has a custom inspector and uses the "Reorderable List",
and the list items also have a custom property drawer.

Currently, the class cannot do anything else.