Create a "Room" ScriptableObject class (unfinished)
This was done to make it possible to associate data with rooms, too.
This is (somewhat) required to make rooms "discoverable", too, which
will be reflected in the rooms list(s) in the game. Additionally,
it also makes it possible to add a custom name and a description
text to all rooms, which are supposed to be shown ingame, as well.

The code is currently not entirely finished. Most of it works
already, but, I wanted to put in some more checks and functionality
to "connect" to the classes from the outside.
Then, I will also have to modify the codes-rooms-mappings class
again.