Finish "Room" class (and custom inspector), add event & comments
The "Room" class should now be pretty finished (hopefully).
It now has documentation and comments, and (public) properties for
all the private fields.
The custom inspector also now displays some warnings when required
data has not been filled out/added yet (like the reference to the
scene).

What's also new is that there is an event now, which is raised
whenever someone changes the "discovery state" of the room. This
means that anyone can listen to this, and be informed whenever the
player discovers that particular room.
(Obviously, this still has to be hooked up to the rest of the game.
It also works in the Inspector, though.)