Recreate the old "KeycodeSceneMapping" class, and finalize it (?)
The old class still used "SceneReference" instances directly. Now,
it of courses uses the "Room" assets - and has been renamed
accordingly.
It also now contains some custom logic for equality comparison, so
that it works the way we want it to work within the "KeycodesList".

The property drawer for the class has also been updated and
extended - it now displays error and warning messages about both
the references (to the keycode and room), and their states (e.g.
whether the room already has a scene file attached, or not).