Add runtime values to the "Keycode" and "Room" class
I totally forgot about this.
These are necessary because changing serialized values of a
ScriptableObject asset otherwise persists after leaving Playmode
(though not if the game is built, only within Unity).
This is usually not the behavior one wants, since the assets would
have to be reset by hand to the intended initial values every time.

Now, during edit time, accessing "isDiscovered" gives you the
"real", serialized variable, but during runtime, you get the
runtime value instead. The runtime value is
automatically once, during deserialization of the asset, when
Playmode starts (includes when the game is built).

There are now also two separate events for listening to the value,
one for the "real" one, for the editor, and one for the runtime
value.
Apart from choosing the correct event to subscribe to, all of this
should be transparent to outside code.

(I hope I did not break anything now. I will see soon...)