Add the model for a simple "Information Board" object
This is a first version, but it's probably good enough to keep it
like that (given how little leftover time we have...).
The model has custom UVs and different materials for each "major"
part of it. It also actually consists of multiple objects, so they
can be independently transformed in Unity.