Add scripts and GameObjects for a scrollable text screen
This already works now (finally). Currently, the pane can only be
moved/scrolled by using a tiny script I made, which just uses
keyboard keys.
The idea is to basically make a "touchscreen" out of it, which works
with Leap Motion. This would give us another Leap Motion
interaction, and would generally be really cool.

If this doesn't work out, I can just add keys to the bottom of the
panel instead (I need keys anyway, for other functionality).

Next up is adding text to the panel via code, and then adding
left- and right-scrolling pages and a "history". Also, the buttons.