Create "Color Preset Library", as an asset, with our PG colors
This can be used to quickly set any color anywhere to one of our
pre-defined "PG colors".
The colors in the library can be edited (and expanded) anytime, by
editing the asset in the Unity Inspector.

Sadly, this does not create _references_ to colors. If we change a
color in the library later, existing colors do not get updated, or
anything like that.