Add some more (very rudimentary and outdated) board shapes for tests
This was to test the size (height) of the boards. Everyone agreed
that it pretty much doesn't matter. All the sizes I had are pretty
much fine.
However, the screens have to be able to be used with your hands
properly, and you need to be able to reach the buttons on the lower
side.
I will probably need to test a bit more with the final model on
Monday.
The font size also isn't that clear. 30 is probably too small, 40
could be fine, 50 feels to big. I'm not sure yet, either.