Rename "IsDiscovered" in "Room" class, replace event with UnityEvent
The variable has been renamed to "HasBeenVisited", because the
"Keycode" class also has a variable called "IsDiscovered", and this
kept confusing people.

I also totally forgot for a while that UnityEvent exists, so I put
C# events into the class instead.
This has been changed now, and there is now a UnityEvent.