Change new "Room" UnityEvent to return the Room instance
Before, it only returned the "Visited" boolean. I didn't know that
the "UnityEvent" class actually does not natively pass on its
"sender", like the native C# events.

This was still fine for just the Room class, but the "KeycodesList"
class actually needs to know the object that the event came from, so
that it can pass it on to its own event.
Also, returning the whole Room instance is more future-proof, since
then we can add other variables or change the types that we fire
events for, etc., and it will still work.