Fix a small bug in "ActivateKeypad"
This bug made it impossible to start the game if the "MasterScene"
isn't loaded.
Although it should normally always be, I didn't like this bug,
especially since there isn't even a clear message for it (you have
to follow the exception Unity throws into the code to understand
what's going on).

Also made the tag comparison going on in the code using the method
meant for it, instead of raw string comparison (which is what Rider
suggests).