Replace auto-rename functionality of "Keycode" asset with a button
This was necessary because there was a critical bug with the
auto-rename going on in the OnEnable() method, which caused Unity to
hang in an infinite loop (and generate gigantic logfiles) on
Playmode start.

Since I'm not entirely sure how to avoid the bug, I've removed the
functionality now. It made doing anything in the custom Inspector of
the class extremely slow, anyway (since it would save the file on
every single change you make).

Instead, there is now a button which does the same as before. When
the button is pressed while the game is "Playing", it will do
nothing.
Hopefully, it is now impossible to break this.