Implement touchscreen controls for the Info Board (a first version)
The information board screen can now by controlled (more precisely,
vertically scrolled) via "touch", by the Leap Motion hands.
This is a bit hard to test at home, but it seems to work just fine.

Currently, this is the simplest possible version, where the code is
just looking for contact with the whole hand.
I will later look a bit into improving this to work with the
fingers.

The code is also still somewhat unfinished, e.g. undocumented and
with debug prints etc.