Add RiggedHands positions to GraspIndicators, fix initial rigged hand swap, ...
Note that previously, there was a delayed HandModel switch on the ModelSwitcher.
This caused the rigged hands to be activated in all the right places, but the debug gesture state was not aware of this.
The current convention is as follows: ALL scripts, components and gameobjects should assume the initial active gesture to be Capsule Hands, aka Debug Mode.
Then, after one second into the game, the Debug Gesture is toggled, such that the Debug mode is off globally, and rigged hands are active.
Background:
It doesn't matter whether debug mode is on or off initially, as long as all scenes and places agree on which debug state is currently active.
As it happens, most scenes started out assuming capsule hands, which is why the convention assumes debug mode initially.
One advantage of having debug mode as the default is that things are visible in the scene view initially, e.g. the body collider objects are visible, whereas with debug mode initially off, they would not be visible at all.