Create a "MessageChannel" asset to transmit text to the Board
This is, of course, based on a ScriptableObject.
The class is really small. The InfoBoard has a reference to the
"channel" asset, and the asset has a (C#) event to listen to and
receive messages when a message is sent "through" the channel. Any
class can take a reference to the asset and then send messages
through it (and in theory, any class could listen to those messages,
then, too).

There's also a corresponding "Message" class, which is also really
tiny.

All the code is undocumented so far.