Work on the "Board Manager" (now "Controller") class
This turned out to be much more work than I thought.
The basic functionality is simple enough, but it is not trivial at
all to just have different "pages" of text (where each page
corresponds to one "message") just slide in and out of the panel.
The stupid "RectTransform" class seems to not play well at all with
being moved around (no matter how you do it, and I tried absolutely
every possible way) in a more complex hierarchy with different
components on different GameObjects meddling with the positioning
and sizing of it, etc.

My old plan was to just move and teleport around the "RectTransform"
with the text and change the text while it's outside of the panel
(and thus invisible), to effectively only have a single GameObject
and reuse that. Because of what I said above, this doesn't work at
all. Also, I realized that I want to already have the next text
(message) sliding in as the old text is still sliding out, so I need
at least two panels already.

I think I know what to do with this now, but it will be even more
effort than I thought, and require more fine-tuning.

Apart from this, most of the functionality is set up now. What's
still missing is finishing the "Lamp Animator" class, and hooking
that up properly to the board.

Yare yare daze.