Create a helper class for turning Board panels into prefabs
I realized that I need to dynamically create Information Board
panels (where each panel corresponds to a single message) at
runtime, so that I can switch them out to sides with an animation,
as was previously already planned.

There might have been other solutions, but this at least works out,
and is pretty flexible, too.

This class was created to give the "Panel" prefabs a definitive
type, and to hold the references to the interesting Components
within the prefab hierarchy, as I do not want to use
"GetComponent()" at runtime all the time.