Add some more important functionality to "TouchscreenControls"
This is necessary because while the Information Board is switching
messages, we cannot use the touchscreen without producing errors,
as it needs a reference to the current Board panel to be able to
scroll that.
Thus, while the Board is switching, it now just turns the
touchscreen functionality completely off, and afterwards turns it
on again.

The Information Board (the "BoardController" class) now also takes
a reference to this script, so that it can switch out the reference
(namely the "ScrollRect") to the newly switched-to panel, so that
the touchscreen works correctly again.

Also, all debug prints have been disabled within the class for now.

Sadly, the class still has no comments or documentation. I want to
test it all out again first (also, time is running out...).