Add a lot of things to Test Scene, rename it to prevent conflicts
This scene now contains the finished board (as a prefab), and I
think I meddled around with some other things, too.

Since someone probably edited the scene on the "master" branch by
now, I renamed it now, so that there are hopefully no merge
conflicts. It is just a test scene, anyway.