Create the main class for creating the User UI
The class, right now, is already capable of creating the panels and
putting them onto the UI canvas (most of the layouting is done by
Unity's UI system, except for settings coming from the UI script).
It also fills the text correctly. Left are the backgrounds, events,
and some other stuff.